requirements.txt
python-dotenv~=1.0.1
mistralai~=1.5.0
pymongo~=4.11
langchain~=0.3.20
langchain-core~=0.3.45
langchain-community>=0.3.21
langchain-mistralai>=0.2.10
llama-index
llama-index-embeddings-mistralai
llama-index-llms-mistralai
llama-index-core
llama-index-utils-workflow
nest-asyncio>=1.5.8
llama-parse>=0.1.0
llama-index-embeddings-mistralai>=0.1.0
nest_asyncio
groq

# ~= means install this version or newer within the same major/minor (e.g. ~=1.5.0 allows 1.5.1, 1.5.9 but not 1.6.0)
